Separate check- CLI commands into distinct modules
- Adjust tests.
- Type hint tests.
